pyproject.toml
[project]
name = "metrics-tracker"
version = "1.0.0"
description = "Work & Individual Metrics Tracker with Streamlit"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "plotly>=5.17.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.7.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[tool.hatch.build.targets.wheel]
packages = []
